Hacks-on-hacks express<->ember routing connected
This commit does some particularly egregious hacks to get the Express
router and the Ember router working together. In particular, though it
ostensibly works, please be careful not to make two requests in parallel
as this will cause the templates to blow up.

This code is not long for this earth, but helps us get a better sense of
what changes are required for a proper solution.