Use public API for creating view element
Previously, we were abusing the internals of the view in order to get
it rendering in Node. As it turns out, there was already a nearly-ready
`createElement` API that we now can use.

In order to get LinkView working, we also had to make URL sanitization
part of the DOM helper. We now override the DOM helper to provide
a Node-supplied method that extracts the protocol from a URL.