Use latest visit API from 2.2 beta
Most of the heavy lifting (runtime stuff) has been moved to
`ember-fastboot-server`. This addon is currently now only responsbile for
building this right app bundle for the fastboot server to consume.

The main integration point between the fastboot server and the app
bundle is the injected `~fastboot/app-factory` module. This allows the
server to create an `Application` instance (not to be confused with an
`ApplicationInstance` instance ...:trollface:) without having to
discover the app name and the app config.

This commit also introduced a convention within the addon: all
(instance) initializers within the `browser` directory will be excluded
from the fastboot app bundle; conversely, the (instance) initializers
contained within the `server` bundle will be removed from the regular
build.