Server 'normal' browser build with --serve-assets
We are producing different content for the server vs browser app bundle
(e.g. the server one has `autoRun` and `autoboot` disabled), so we need
to serve the browser version with `--serve-assets`, otherwise the app
won't boot in the browser.

The `--assets-path` is somewhat of a quick fix. We should probably run
the build process twice and store them in the same location but with a
different suffix, like `dummy.fastboot.js` and `index.fastboot.html`.