Add fastboot-config with whitelisted modules
This commit adds generation of a `fastboot-config.json` to the
`fastboot-dist`. For now, this only includes one field which is a list
of npm packages that are needed when the app is running in FastBoot.

For example, a `fastboot-config.json` may look like this:

```json
{
  "moduleWhitelist": ["najax", "node-fetch"]
}
```

The list of whitelisted modules is derived from the app's installed
addons. Addons can add a `fastBootDependencies` key to their
`package.json`'s `ember-addon` field:

```json
{
  "name": "ember-network",
  "version": "0.1.0",
  "ember-addon": {
    "fastBootDependencies": ["node-fetch"]
  }
}
```

Each addon's list of dependencies will be aggregated and placed into
a single list of whitelisted modules.