Replace ajax{Success,Error} with an override of `adapter._ajaxRequest`
Per @tchak's suggestion, moving the jQuery request in the Ember Data
adapter [here][ed-pr], the code to replace `jQuery.ajax` with `najax`
becomes trivial, and replaces all code duplication.

There was an update to najax that is needed for Ember data to be able to
parse the headers. That code is available on the master branch of najax,
but has not been released.

This also depends on the latest beta branch of Ember Data, and will be
compatible with ED 2.4

[ed-pr]: https://github.com/emberjs/data/pull/4113