Adds hot reloading/zero downtime deploy
This commit adds support for reloading a running FastBoot server.

In the programmatic case, you can call `server.reload()` to have it
destroy the old app and re-read the new version off the disk.

If running from the command line, you can send SIGUSR1 to the process
and it will reload the app from disk.