Use fastbootAppConfig.htmlFile if it is present, default to 'index.html'
This allows a developer to specify a custom htmlFile that Fastboot will
use instead of index.html as its target for injecting rendered head or
body content. This is useful when your app needs to be rendered in a
wholly different HTML shell, such as when using Ember to generate a
valid AMP page (AMP requires a certain HTML structure for its pages and
the default index.html in ember-cli does not fit this).
The custom htmlFile should be a standard HTML file (no handlebars usage) in
the app's `public/` directory and needs to set its own
`<!-- EMBER_CLI_FASTBOOT_HEAD -->` and `<!-- EMBER_CLI_FASTBOOT_BODY -->`
placeholders so that fastboot injects the rendered content into it correctly.