Better shoebox API
This adds a new higher-level API for the shoebox.

```js
  model() {
    let fastboot = this.get('fastboot');
    return RSVP.hash({
      // In this example, we generate a random number once on the server,
      // and then the  client app will receive the same number.
      random: fastboot.withShoebox('my-random-number', () => Math.random())
    });
  }
```

`withShoebox` acts as a memoization service. On the server, the given callback will run and generate a value, which will be saved in the shoebox and then returned. On the client, the callback will not run if the value is already present in the shoebox.

`withShoebox` returns a promise, because the callback is allowed to return a promise.