Ensure 404's receive the proper response code.
Prior to this change, requesting a URL that did not exist in the
app resulted in a 500 response code (which is not correct).

When we encounter an unrecognized URL we should simply call `next`
and allow any downstream middlewares to handle the request. If
nothing handles the request, Express will set the status code
to 404 on its own.