Fix tests
It is unknown right now as to why these are suddenly lowercase, this
directly reverts a change from about a year ago by Stef.

https://github.com/ember-fastboot/ember-cli-fastboot/commit/67102c964e6b276e3f318f01eca56944fe337e3c#diff-c2a8324c5e490c750018e7ed31eb42e9

Previously somewhere in our middleware stack upcased the `UTF-8` portion
of the Content Type header this doesn't seem to be happening anymore.

This fix seems innocuous for now.  Perhaps worth investigation later.