Use case-insensitive matching when comparing content-type
See #583

The "charset=utf-8" in content-type sometimes changes in tests depending
on what express middleware serves the response.

The charset value is case-insensitive according to IANA (see:
https://www.iana.org/assignments/character-sets/character-sets.xhtml)

This adds chai-string (http://chaijs.com/plugins/chai-string/) as a devDep and
uses its `equalIgnoreCase` matcher in tests that check the value of
the content-type header.