Dev server forwards out of scope requests
When running a dev server and an app is mounted at a root URL other than `/`, it is possible for requests to make it to the fastboot server that do not have the `x-broccoli` header.  Moving parsing of this header deeper into the request allows for these out of scope requests to be forwarded on.