Refactor sandboxGlobals -> buildSandboxGlobals
This changes the system from providing default set of shared (and
therefore mutable) global properties to using a builder function to
generate the set of globals to be used _per visit_.

The `buildSandboxGlobals` function will receive the default set of
globals that FastBoot creates (currently this is `najax` and
`FastBoot`), and whatever the `buildSandboxGlobals` function returns is
what will ultimately be used. If `buildSandboxGlobals` is not provided,
a default implementation is used (it is essentially `defaultGlobals =>
defaultGlobals;`).

For example, to specify a custom global property named `AwesomeThing` to
be accessed within the sandboxed context:

```js
let fastboot = new FastBoot({
  distPath: 'some/path/here',
  buildSandboxGlobals(globals) {
    return Object.assign({}, globals, {
      AwesomeThing: 'Taco Cat'
    });
  }
});
```

If `sandboxGlobals` is passed (and `buildSandboxGlobals` is not) issue a
deprecation and automatically create a `buildSandboxGlobals` of the
following:

```js
globals => Object.assign({}, globals, options.sandboxGlobals);
```