Remove ember-source from devDependencies.
We were using it primarily to test that our
`FastBootHeaders.prototype.unknownProperty` method functioned
properly. This refactors things to use a simulated `Ember.get` to allow
the same test scenario to succeeed, but avoid having to depend on Ember
itself.

Note: this does not remove any "real" testing of various Ember versions,
most of our tests are using fixtures with a fully compiled ember-cli
applications.