implement a new html-oriented manifest format
This implements a new v5 schema for fastboot. Unlike the previous json-oriented formats, this is an html-oriented format. HTML is better understood by general purpose build tools, which makes it easier to integrate with embroider. It also allows us to eliminate some redundancies -- for example, the app config no longer needs to exist in both package.json and the meta tags, since the HTML oriented format can use it directly from the meta tags.

In package.json, the v5 schema supports:

- `fastboot.schemaVersion: 5`
- `fastboot.moduleWhitelist: []`: unchanged meaning from v4.
- `fastboot.htmlEntrypoint: 'index.html'`: the path to your HTML entrypoint file
- `name`: note that in the new format, we take the appName directly from the package.json top level. This field is used for nothing else by the time we get to fastboot, so there was no reason for redundancy.

The HTML entrypoint file is any valid HTML, plus these fastboot-specific extensions:

- `<fastboot-script>` tags work like `<script>` tags, but they will only run in fastboot and be stripped out before serving to browsers.
- `<script data-fastboot-ignore></script>` will *not* run in fastboot, and the data attribute will be stripped out before serving to browsers.
- `<script src="https://some-cdn.com/app.js" data-fastboot-src="assets/app.js"></script>` tells fastboot the local path to an asset that may have a different public URL for browser consumption. The data-fastboot-src attribute is stripped before serving to browsers.

New html-oriented manifest format for embroider

implementing data-fastboot-ignore and data-fastboot-src

making htmlEntrypoint the new v5 schema

And refactoring so we truly hide the differences in schema versions from the rest of the program.

schema v5

This makes the htmlEntrypoint format be the official v5 schema.