Prevent errors when `fastboot-body-end` is not present.
The current implementation of `clear-double-render` guards for when
`fastboot-body-start` is not present, but does not check if
`fastboot-body-end` is present before attempting to remove it (therefore
causing an `Cannot call 'parentElement' of null` error).

Note: There are still some issues with this approach (e.g. why is
`fastboot-body-start` present but `fastboot-body-end` is missing), and
this change does not directly address all of the issues. It **does**
prevent a masking error from happening.